+++ app/src/main/java/com/example/movieapp/data/model/product/SizeProductModel.kt
@@ -4,6 +4,6 @@ import java.io.Serializable
 
 data class SizeProductModel(
     val size: Int = 0,
-    var amountSize: Int = 0,
+    var amountSize: Long = 0L,
     var isSelected: Boolean = false
 ): Serializable
